Make volume normalization a separate argument
